+++ prepare.bash
@@ -13,7 +13,12 @@
 # You should have received a copy of the GNU General Public License
 # along with Onur. If not, see <https://www.gnu.org/licenses/>.
 
-mkdir -pv ~/.config/onur
-cp -v ./examples/* ~/.config/onur
+mkdir -p ~/.config/onur
+
+cp ./examples/* ~/.config/onur
+
 touch ~/.config/onur/emptyfile.json
-ln -sf ~/nonexistentfile ~/.config/onur/baz.json
+
+ln -sf ~/nonexistentfile ~/.config/onur/danglingsymlink.json
+
+rm ~/.config/onur/lua.json && cp ./examples/lua.json ~/ && ln -sf ~/lua.json ~/.config/onur/lualink.json